remove unnecessary useEffect hook
